Revert "Manager openshift update"
This reverts commit b6c10d7a786c2476693f3677f2e128aed348ba3e.

Manager openshift update

This is the manager of the other scripts, this
one takes care of:
- user input via web application, using flask
-
communication with the other scripts using rabbitcoat
- displaying the
results to the user

The last part wasn't yet tested, as there are no
results for now.

As a side note, for now I think the entry point of the
application is:
- wsgi.py by default
- or what's specified in the
README.md file inside '''bash'''
Not sure about this, but couldn't find
another rule.

NOTE: Configuration should be updated with credentials before running